sin solucionar entitas cache

+++ examples/MatchOne/src/main/java/com/ilargia/games/components/Position.java
@@ -5,10 +5,10 @@ import com.ilargia.games.entitas.codeGenerator.Component;
 
 @Component(pools = {"Game"})
 public class Position implements IComponent {
-    public float x;
-    public float y;
+    public int x;
+    public int y;
 
-    public Position(float x, float y) {
+    public Position(int x, int y) {
         this.x = x;
         this.y = y;
     }
